[MODEXPS-42] - Remove EDIFACT
